Copy pycompwa to install location
- Add pytest to travisCI
