+++ pycompwa/README.rst
@@ -1,0 +1,1 @@
+# pycompwa - ComPWA python module
