+++ pycompwa/pycompwa/tests/EpEmToChicPipPim.py
@@ -1,6 +1,6 @@
 import logging
 
-from expertsystem.ui.system_control import (
+from pycompwa.expertsystem.ui.system_control import (
     StateTransitionManager, InteractionTypes)
 
 logging.basicConfig(level=logging.INFO)
