+++ pycompwa/pycompwa/tests/JPsiToPi0PipPim.py
@@ -3,10 +3,10 @@
 """
 import logging
 
-from expertsystem.ui.system_control import (
+from pycompwa.expertsystem.ui.system_control import (
     StateTransitionManager, InteractionTypes)
 
-from expertsystem.amplitude.helicitydecay import (
+from pycompwa.expertsystem.amplitude.helicitydecay import (
     HelicityDecayAmplitudeGeneratorXML)
 
 logging.basicConfig(level=logging.INFO)
