+++ pycompwa/pycompwa/tests/test_D0ToKsKpKm.py
@@ -3,9 +3,9 @@
 """
 import logging
 
-from expertsystem.ui.system_control import StateTransitionManager
+from pycompwa.expertsystem.ui.system_control import StateTransitionManager
 
-from expertsystem.amplitude.helicitydecay import (
+from pycompwa.expertsystem.amplitude.helicitydecay import (
     HelicityDecayAmplitudeGeneratorXML)
