+++ pycompwa/pycompwa/tests/unittests/topology-builder/test_simplestatebuilder.py
@@ -1,5 +1,5 @@
-from expertsystem.topology.graph import InteractionNode
-from expertsystem.topology.topologybuilder import (
+from pycompwa.expertsystem.topology.graph import InteractionNode
+from pycompwa.expertsystem.topology.topologybuilder import (
     SimpleStateTransitionTopologyBuilder)
